refactor: 가격 필드 Int 형으로 변경 - 부동 소수점은 연산시에 오차가 생기기 때문에 정수형으로 사용 - 13.33$ 와 같이 소수점이 필요한 경우 단위를 센트로 사용(1333)

+++ src/main/kotlin/com/chaw/shopping_note/app/receipt/application/dto/GetReceiptResponseDto.kt
@@ -13,9 +13,9 @@ data class ReceiptItemResponseDto(
     val id: Long,
     val productName: String,
     val productCode: String,
-    val unitPrice: Double,
+    val unitPrice: Int,
     val quantity: Int,
-    val totalPrice: Double,
+    val totalPrice: Int,
     val category: Category,
     val updatedAt: LocalDateTime,
 )
@@ -25,7 +25,7 @@ data class GetReceiptResponseDto(
     val store: StoreResponseDto?,
     val items: List<ReceiptItemResponseDto>,
     val purchaseAt: LocalDateTime,
-    val totalPrice: Double,
+    val totalPrice: Int,
     val totalCount: Int,
     val createdAt: LocalDateTime,
 ) {
@@ -49,15 +49,15 @@ data class GetReceiptResponseDto(
                         id = it.id!!,
                         productName = it.productName,
                         productCode = it.productCode,
-                        unitPrice = it.unitPrice.toDouble(),  // 타입 맞추기
+                        unitPrice = it.unitPrice,
                         quantity = it.quantity,
-                        totalPrice = it.totalPrice.toDouble(),
+                        totalPrice = it.totalPrice,
                         category = it.category,
                         updatedAt = it.updatedAt!!
                     )
                 },
                 purchaseAt = receipt.purchaseAt!!,
-                totalPrice = receipt.totalPrice.toDouble(),
+                totalPrice = receipt.totalPrice,
                 totalCount = receipt.totalCount,
                 createdAt = receipt.createdAt!!
             )
